move concept graph to torch_concept level
